src/tavily-search/pyproject.toml
[project]
name = "tavily-search"
version = "0.1.0"
description = "A MCP server for AI-powered web search using Tavily API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0",
]

[[project.authors]]
name = "Your Name"
email = "your.email@example.com"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
tavily-search = "tavily_search:main"
